optimazion

+++ src/http_func.c
@@ -157,7 +157,7 @@ int http_func(const char *fullurl, char *buffer, int num, int method,const char 
 		if(post_string == NULL){
 			snprintf(send_buffer,8192,"%s %s%s HTTP/1.0\r\nHost: %s\r\nUser-Agent: Mozilla/5.0\r\n\r\n",mt,u->path,u->query,u->hostname);
 		} else {
-			snprintf(send_buffer,8192,"%s %s%s HTTP/1.0\r\nHost: %s\r\nUser-Agent: Mozilla/5.0\r\nContent-Type: application/x-www-form-urlencoded\r\nContent-Length: %d\r\n\r\n%s",mt,u->path,u->query,u->hostname,(int)strlen(post_string),post_string);
+			snprintf(send_buffer,8192,"%s %s%s HTTP/1.0\r\nHost: %s\r\nUser-Agent: Mozilla/5.0\r\nContent-Type: application/x-www-form-urlencoded\r\nContent-Length: %zu\r\n\r\n%s",mt,u->path,u->query,u->hostname,strlen(post_string),post_string);
 		}
 
 #if DEBUG > 0
@@ -222,7 +222,7 @@ int http_func(const char *fullurl, char *buffer, int num, int method,const char 
 			if(post_string == NULL){
 				sprintf(send_buffer,"%s %s%s HTTP/1.0\r\nHost: %s\r\nUser-Agent: Mozilla/5.0\r\n\r\n",mt,u->path,u->query,u->hostname);
 			} else {
-				sprintf(send_buffer,"%s %s%s HTTP/1.0\r\nHost: %s\r\nUser-Agent: Mozilla/5.0\r\nContent-Type: application/x-www-form-urlencoded\r\nContent-Length: %d\r\n\r\n%s",mt,u->path,u->query,u->hostname,(int)strlen(post_string),post_string);
+				sprintf(send_buffer,"%s %s%s HTTP/1.0\r\nHost: %s\r\nUser-Agent: Mozilla/5.0\r\nContent-Type: application/x-www-form-urlencoded\r\nContent-Length: %zu\r\n\r\n%s",mt,u->path,u->query,u->hostname,strlen(post_string),post_string);
 			}
 #if DEBUG > 0
 			printf("%s[DEBUG]%s[%s][%s] Connected with %s encryption\n", REDBOLD,NOCOLOR,__FILE__,__func__, SSL_get_cipher(ssl));
